Remove support for the deleted bit in the cell description
The previous idea of marking cells as deleted and compacting the
range table independently was discarded as problems with the binary
cell search arose. This commit therefore removes support for the
deleted bit in the cell description and the syscall for compacting
the range table. The compaction now happens everytime a cell is
deleted.